NotificationUsageAutoAddProcessor と API/DTO 定義を別ファイルに分離
